requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.29.3
setuptools==75.8.1
